BUG: interactive test for scaling seed widgets 2d vs 3d viewers
When using a vtkSeedWidget in a 2d slice view, the correct
camera has to have it's position set to mirror the viewing
distance in the related 3d window.

To run, in the CTK-build dir:
ctest -R ctkVTKSliceAndRenderViewWithSeedWidgetTest -VV
Copy the command line and paste it with the addition of -I,
click in the top window, press 'i' to enter interactive placement
mode and place a seed. The top window is a 2d slice viewer,
the bottom window is a 3d render viewer. A second seed widget is
created to mirror the first one.

Slicer Issue #3547