ENH: add ability to delete individual series, studies, patients
Add to the browser Remove button the information that it will
delete all selected series, studies, patients.
Add support for right click custom context menus in the
patient, study, series tables. Translate the local points
to global ones in terms of the table and signal
right clicked with the global position so that the browser
can position the context menu appropriately.
Add the context menu to the DICOM browser to allow
deleting at just one level.
Added a test for the DICOM browser, with the ability
to open it in interactive mode to test the right clicks.

Note: left the response to double click on the series for Slicer
testing as the right click is not being detected within
that context. More testing is needed.

Slicer bug:
http://www.na-mic.org/Bug/view.php?id=3792

Issue #3792