BUG: move right click to table view
When Slicer makes a custom DICOM browser, it moves the table views
directly into the new window, extracting them from the DICOM
table. The previous commit on this branch had set up a right click
signal propagation assuming that the whole table widget was present
(with search box and label), but with just the database view moved,
the signals needed to be shifted so that they would be connected
directly to the table manager in a useful manner.

The table view takes care of mapping the point to global, the table
manager propagates the signal, and the browser responds to it with a
table specific context menu.

Took out the extra double click that was there for testing.

Added selection info to the context menu. Using UIDs for now, though
the UIDs are hidden in the table so this can be changed, it's intended
as a double check before deleting, in conjunction with the cancel
option.

Updated the text on the delete actions to emphasise that it's
selected items that will be deleted.

Update the DICOM table manager to translate the custom context menu
requested signal to the table specific XRightClicked signal, removing
the map to global at that level.

In the table view, set the context menu policy and connect the
custom context menu requested signal to a local slot that does the
mapping of the point to the global widget space so that the browser
can pop up the context menu at the right location.

Slicer issue #3792